added README to kinome -mutants, cleaned up duplicated files in the construct directory
