+++ constructs/96-kinase-constructs/PDB-constructs/alignments/.gitignore
@@ -1,0 +1,2 @@
+*.html
+*.css
