--- constructs/PDB-constructs/.gitignore
@@ -1,1 +1,0 @@
-DLPold
