--- constructs/PDB-constructs/alignments/.gitignore
@@ -1,2 +1,0 @@
-*.html
-*.css
